### 2016-12-21 #144 @arfon
@arokem - would you be willing to edit this submission for JOSS?
