### 2017-01-04 #144 @arfon
> @arokem - would you be willing to edit this submission for JOSS?

Friendly new year reminder @arokem 😁  - this looks closest to your editorial topic areas.
