### 2017-01-05 #144 @Kevin-Mattheus-Moerman
@jcfr @aylward if you are not involved in this project, are you able to review (or suggest reviewers for) this submission for http://joss.theoj.org/?
