### 2017-01-06 #144 @fedorov
@Kevin-Mattheus-Moerman can you please point me to your reviewer guidelines? I could not quickly locate them on http://joss.theoj.org/.
