### 2017-01-06 #144 @lassoan
I would be happy to review, but not sure what or how. I only see a repository with some code and doc. Is there a rendered document somewhere? Should I review and/or test the code?
