### 2017-01-06 #144 @Kevin-Mattheus-Moerman
@fedorov @lassoan the reviewer guidelines are found here http://joss.theoj.org/about#reviewer_guidelines

The review process focuses largely on the claims about the software and the software itself. The paper is a very short document since our view is that properly documented software forms the more important part of the software publication.

A paper PDF is  generated based on the paper.md file which for this project is found here:
https://github.com/KitwareMedical/SlicerITKUltrasound/blob/master/paper.md

The paper text is also rendered here:
http://joss.theoj.org/papers/746880f9243feff57502fae59abce923

This is an example of an accepted paper:
http://joss.theoj.org/papers/73fcd8f3fcc262db670ec8fccfdad706

You can view the review process that occurred for that paper here:
https://github.com/openjournals/joss-reviews/issues/43

As you can see, once a reviewer is assigned a check list appears which forms the basis of the review process.

Let me know if you have any other questions.

Kevin
