### 2017-01-06 #144 @Kevin-Mattheus-Moerman
@fedorov our quasi-automated system can currently handle 1 reviewer which is currently set to @lassoan. Nevertheless I'd like you @fedorov to act as the second reviewer. You can leave comments here for the authors to act on and also follow the checklist that will appear here shortly.
