### 2017-01-06 #144 @lassoan
The submission fulfills the JOSS requirements.

My only concern is that the software's scope (as described in the title) is "ultrasound image formation, processing, and analysis", but now it only addresses a small task in ultrasound image formation (scan conversion). Many more modules will be needed to make the software fulfill its promise.
