### 2017-01-06 #153 @whedon
[REVIEW]: SlicerITKUltrasound: A 3D Slicer extension for ultrasound image formation, processing, and analysis
**Submitting author:** @thewtex (<a href="http://orcid.org/0000-0001-9475-3756">Matthew McCormick</a>)
**Repository:** <a href="https://github.com/KitwareMedical/SlicerITKUltrasound" target ="_blank">https://github.com/KitwareMedical/SlicerITKUltrasound</a>
**Version:** v0.2.0
