### 2017-01-06 #144 @arfon
@Kevin-Mattheus-Moerman - #153 is the main review issue so we should move the conversation over there.
