### 2017-01-06 #153 @Kevin-Mattheus-Moerman
@lassoan @fedorov @thewtex apologies for the confusion. This is where the review process will happen. Reviewers, see reviewer instructions here http://joss.theoj.org/about#reviewer_guidelines and follow the above checklist. Let me know if you have any questions.
