### 2017-01-16 42e83d5
DOC: Clarify that the extension currently performs scan conversion
The extension is currently used for scan conversion of pre-computed ARFI
images, etc. while image formation may be integrated in the future. Improve
the title and description so this is more clear.
