### 2017-01-16 #153 @thewtex
```
The submission fulfills the JOSS requirements.

My only concern is that the software's scope (as described in the title) is "ultrasound image formation, processing, and analysis", but now it only addresses a small task in ultrasound image formation (scan conversion). Many more modules will be needed to make the software fulfill its promise.
```

@lassoan Thank you for the review.  Yes, this could be clarified.  Please review this patch:

  https://github.com/KitwareMedical/SlicerITKUltrasound/pull/49

that changes the title and description to clarify that the extension performs scan conversion on image formation that is currently executed elsewhere.
