### 2017-01-16 #153 @thewtex
@Kevin-Mattheus-Moerman @arfon Thanks for moving this forward!

Examining the preview here:

  http://joss.theoj.org/papers/746880f9243feff57502fae59abce923

The references in the *paper.bib* file are not rendered. Is this intended? Is there a problem with the *paper.md* file?
