### 2017-01-16 #153 @arfon
> The references in the paper.bib file are not rendered. Is this intended? Is there a problem with the paper.md file?

@thewtex here's the compiled paper PDF [10.21105.joss.00153.pdf](https://github.com/openjournals/joss-reviews/files/709556/10.21105.joss.00153.pdf)
