### 2017-01-16 #153 @lassoan
@thewtex Thanks for updating the text to make things a bit more clear.
