### 2017-01-16 #153 @lassoan
I think I kind of understand the goal of this journal and using github for storing data could work. However, you should really improve the experience for reviewers. The process and usage of tools should be much simpler, more convenient and similar to existing review systems. It's also unclear for me how useful an initial, one-time review of a software repository can be, because a software library is continuously changing in time.
