### 2017-01-17 #153 @Kevin-Mattheus-Moerman
@lassoan
> However, you should really improve the experience for reviewers.
> The process and usage of tools should be much simpler, more convenient and similar to existing review systems.

Thanks for your feedback. We are continuously trying to improve our system and also our review process.  Feel free to suggest improvements to our system. You can add comments to our open issues or create new ones.
>  The process and usage of tools should be much simpler, more convenient and similar to existing review systems.

What existing systems are you referring to? Current software journals? Traditional journals? We deviate from other journals since we uniquely focus on the software. Let us know if you have suggestions for improvement.
> It's also unclear for me how useful an initial, one-time review of a software repository can be, because a software library is continuously changing in time.

I understand your point. JOSS came about to help provide academic credit for software, i.e. to create journal publication style credit, which involves having formal peer review, and upon acceptance assigning CrossRef DOI's enabling citations which "count". See also a blogpost from the JOSS lauch http://www.arfon.org/announcing-the-journal-of-open-source-software
In relation to future versions we feel that major releases can be resubmitted for review and can be a candidate for a new paper. This enables continuous review and accurate citing of the correct releases.
