### 2017-01-17 #153 @fedorov
> JOSS came about to help provide academic credit for software

FYI, at least in my field, a publication in a journal that is not indexed by Medline is not recognized, so publication itself gives no credit. It is more conventional to publish a software article in a Medline-indexed journal so that both academic credit for the publication and citations can come from the same location. I am sorry to say this, but as is, I see no benefit in submitting to JOSS. For academics, the criteria for choosing a journal is typically Medline-indexed (hard requirement) and journal impact factor. Do you do any work to advance JOSS in these two areas?
