### 2017-01-17 #153 @Kevin-Mattheus-Moerman
Thanks @fedorov @lassoan. These are good comments. To avoid clogging this review issue I invite you to look at / open issues on these questions here: https://github.com/openjournals/joss

Also some of the issues with MEDLINE/Pubmed indexing are addressed here:
https://github.com/openjournals/joss/issues/153
