### 2017-01-18 #153 @fedorov
@Kevin-Mattheus-Moerman I agree, this kind of discussion does not belong to this specific issue. Sorry for hi-jacking the topic.
