### 2017-01-18 #153 @Kevin-Mattheus-Moerman
@fedorov that is fine. We are a novel journal and there are definitely several creases we need to iron out and some of those are highlighted first by the reviewers at this stage so it is fine.

Now on to the review process here.
@lassoan have you finalised your end of the review process or or will you provide more detailed feedback?
@fedorov @thewtex can you confirm you are also reviewing this submission?

Let me know if you have any questions.
