### 2017-01-18 #153 @lassoan
I'm done with my review at this stage of the project.
