### 2017-01-18 #153 @fedorov
I guess I am still confused by the review process. There is a bunch of checkboxes on top of this issue. Who is supposed to check all of those? Is that part of the review process or not?
