### 2017-01-18 #153 @Kevin-Mattheus-Moerman
@fedorov One of the current flaws in our system is that only 1 reviewer (currently @lassoan) can be assigned to this issue through our system to check those boxes. Handling multiple reviewers in that sense is a work in progress. Of course multiple reviewers are desired so in order to handle multiple we need to be a bit creative at the moment. Currently only @lassoan is able to tick those boxes. However we will not converge on a decision to publish until we have reasonable agreement between all reviewers that all boxes can be ticked and that this submission is deemed acceptable. For the moment the other reviewers can post any review comments they have here. Also feel free to quote portions of the check-box text to comment specifically on the points highlighted there.

Here is an example of a completed review process:
https://github.com/openjournals/joss-reviews/issues/64

and here is an example of a lengthy review process with multiple reviewers working their comments in:
https://github.com/openjournals/joss-reviews/issues/47

Hope that helps.
