### 2017-01-19 KitwareMedical/SlicerITKUltrasound#50 @fedorov
Sample dataset would be appropriate
In connection to https://github.com/openjournals/joss-reviews/issues/153, I could not find any dataset that would allow to demonstrate the usage of the software.
