### 2017-01-20 97bf90b
DOC: Add links to example input data
Issue #50
