+++ Documentation/source/Modules/ScanConversion/CurvilinearArray.rst
@@ -10,4 +10,5 @@
   converted.
 
   This output can be reproduced by running the ``ScanConvertCurvilinearArrayTest``
-  test in the project repository.
+  test in the project repository. Example input data can be found `here
+  <http://slicer.kitware.com/midas3/folder/3646>`_.
