### 2017-01-20 KitwareMedical/SlicerITKUltrasound#50 @thewtex
@fedorov Good point. Please review this PR: https://github.com/KitwareMedical/SlicerITKUltrasound/pull/51
