### 2017-02-10 #153 @fedorov
I went over the items on the check list, and I am fine with all of them except:

- [ ] Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support - **contributing guidelines are missing**
- [ ] References: Do all archival references that should have a DOI list one (e.g. papers, datasets, software)? - **references are missing in [paper.md](https://github.com/KitwareMedical/SlicerITKUltrasound/blob/master/paper.md) (unless they are somehow smartly generated in some other view from .bib file!) and no DOIs are listed in the .bib file**

Over to @thewtex.
