### 2017-02-10 #153 @Kevin-Mattheus-Moerman
Thanks @fedorov, In relation to the second point, we do generate a PDF with references, here is the compiled paper PDF with some references and DOI's in place: [10.21105.joss.00153.pdf](https://github.com/openjournals/joss-reviews/files/709556/10.21105.joss.00153.pdf), so please review that document.

@thewtex let me know when you've addressed these comments.
