### 2017-02-10 #153 @fedorov
@Kevin-Mattheus-Moerman indeed this resolves my second concern. Thanks!
