### 2017-02-10 #153 @fedorov
@thewtex it may be my personal preference, and of course independent of this paper review, but I would recommend using a different color scheme for the documentation: light blue and grey do not work well together, in my opinion.

![image](https://cloud.githubusercontent.com/assets/313942/22841006/c811de90-ef9d-11e6-9f16-a4bdbfa25551.png)
