+++ CONTRIBUTING.md
@@ -1,0 +1,14 @@
+# Contributing to SlicerITKUltrasound
+
+Contributions are welcome and encouraged.
+
+To contribute, create a pull request on the [project's GitHub
+repository](https://github.com/KitwareMedical/SlicerITKUltrasound).
+Respectful and collaborative discourse is appreciated and expected.
+Please patiently address all automated builds and review comments.
+
+For more information how to use the Git version control system, see the
+excellent [ProGit
+book](https://git-scm.com/book/en/v2/Getting-Started-Git-Basics).  This
+repository uses common GitHub practices, sometimes referred to as the [GitHub
+flow](https://help.github.com/articles/github-flow/).
