### 2017-02-10 #153 @thewtex
@fedorov Thanks for the review.

I've added a basic *CONTRIBUTING.md* here: https://github.com/KitwareMedical/SlicerITKUltrasound/pull/52

Agreed on the colorscheme, too. I've removed the bright blue and used black for the normal text.
