### 2017-02-10 #153 @fedorov
I don't have any further concerns about the check list at the top of this review. Looks good to me to be accepted!
