### 2017-02-10 #153 @Kevin-Mattheus-Moerman
@thewtex perhaps you have done this already, but can you create an updated archive of the reviewed software in Zenodo (or figshare/other service) and update this thread with the DOI of that archive?
