### 2017-02-13 #153 @arfon
@lassoan @fedorov many thanks for your review here and to @Kevin-Mattheus-Moerman for editing this one ✨

@thewtex your paper is now accepted into JOSS and your paper DOI is http://dx.doi.org/10.21105/joss.00153 ⚡️ 🚀 💥
